Adding some useful comments in SignallingProtocol. SignallingProtocol has nothing to do with web sockets so renamed its folder to Protocol
