+++ Common/src/pixelstreamingcommon.ts
@@ -1,7 +1,7 @@
 export { Logger } from './Logger/Logger';
 export { ITransport } from './Transport/ITransport';
 export { WebSocketTransport } from './Transport/WebSocketTransport';
-export { SignallingProtocol } from './WebSockets/SignallingProtocol';
+export { SignallingProtocol } from './Protocol/SignallingProtocol';
 export { IMessageType } from "@protobuf-ts/runtime";
 export { BaseMessage } from './Messages/base_message';
 export { MessageRegistry } from './Messages/message_registry';
